Robust versioning for our packages
It's fine for the builder to say "use this tag/branch", but our build
system needs to translate that into a proper version so that the OpenWRT
build suite doesn't thing "foo-master.tar.xz" is always up-to-date.

We'' be using foo-<branch|tag>-<commitid>.tar.xz to solve this problem.

MFW-80 @exclude