Add crash-collector
openwrt includes a kernel patch to store kernel crash logs in memory to be
collected on the subsequent boot.  This is kind of hacky as it depends on
nothing touching that particular memory during a reboot.  However, it seems
to work on the devices we care about (wrt1900, wrt3200, omnia), so the
crash collector will grab crash logs if they exist, and put them in
/etc/config/crash for future analysis.  To make sure we don't fill up the
storage part it currently only keeps the latest 5 crash logs.  This costs
about 25k.  We can revisit the storage policy later if we need to.