Use actual timestamp of source files when building images
This breaks build reproducibility, which is unfortunate. However, go-gin
doesn't support ETag handling, which would be the proper solution to the
problem of the UI not reloading.

Do not expect this change to make *all* files in the image show the
latest build timestamp: since we do incremental builds (re-using all the
existing clones from one build to the next), source files only get
updated if there are new git commits that actually modified them between
the previous build and the current one.

MFW-227 @exclude