container: Don't build for linksys wrt32x
The firmware layout for the wrt32x only reserves
3MB for the kernel binary.  The kernel configs necessary
for container support produce a 3.2MB kernel.  For now,
disable the container package when building for the
wrt32x.  In the future, we can investigate slimming down
the container enabled kernel enough to fit into 3MB, and
then remove this hack.

MFW-315