caswell-rssi-leds: Reset modem at boot
This will ensure we are starting from a fresh configuration.  In the
cases of a loss of power or a sysupgrade we may not have properly
shut down the LTE connection prior to rebooting.  That can leave
the modem in a state in which it thinks its connected, but isn't.
In this state, no amount of configuring the device will ever get
it back on line.  Only a modem reset will do.

MFW-802