caswell-rssi-leds: Disable gnss at boot
In order to get Verizon (and possibly others) certification we need
to ensure the GNSS Location Service feature of our modem is disabled.

Disabling it should persist even through reboots, but just in case
we'll just send the disable command every time we boot.  The only
way to send this command is via the modems serial connection with
AT commands, so enable the kernel configs necessary expose the serial
port to linux.

MFW-887