feeds.conf.mfw: Point back to master
Every time we create a release branch of mfw_build and change this
file to point to the release branch of mfw_feeds, that change gets
auto merged into the master branch of mfw_build, necessitating a
manual change to point it back to master.  Is there a better way?