git-remote-describe: use ls-remote instead of cloning locally
In the grand scheme of things, cloning locally before using "describe
--always" yields more readable objects, especially when they refer to
recent tags.

However, classd is *extremely* costly to clone (a couple of minutes on
average), and that's a huge painpoint when building clean.

When we also consider the fact that there has not been *one* annotated
tag in any of the "satellite" repositories (those referenced in
mfw_feeds) since MFW started as a product, we can conclude that a bare
commit hash seems to be enough to identify versions of components in our
feeds. Therefore ls-remote provides a more efficient implementation for
retrieving those versions.