runtests: fix building from scratch
Removing the runtime dependency on python3-jsonschema had a subtle
side-effect: on a clean host, with no prior MFW builds, runtests
now fails to build because of a setuptools-related issue in
the build of one of its runtime dependency, namely python3-pytest.

Since pytest is not directly used by runtests, removing it is the low
hanging fruit, but it doesn't yet fully explain why python3-pytest
builds fine when python3-jsonschema is taken out of the dependency
chain in another MFW leaf package: I'm still working on root-causing
this, but in the meantime this commit should fix the overall MFW build
issue on Jenkins.

BTW, this "build from scratch" issue doesn't show up in barney-based
builds, because `sync-settings/barney.yaml`, which provides the
`mfw_pkg/sync-settings` target, is not using
`code.arista.io/mfw/build%toolchain as its base`; in other words, the
way we build and declare build-dependencies for sync-settings
is *wildly* different between Jenkins and barney:

  - Jenkins: regular openwrt-based, using
    `mfw_feeds/sync-settings/Makefile` for all declarations, in the
    standard bullseye build container

  - barney: barnzilla-based, with build dependencies declared directly
    in barney.yaml